+++ gamemodes/terrortown/gamemode/client/cl_help/populate_legacy.lua
@@ -119,3 +119,13 @@ HELPSCRN.subPopulate["ttt2_legacy"] = function(helpData, id)
 		end)
 	end
 end
+
+---
+-- This hook was used in TTT to populate the F1 settings menu. It shouldn't be used
+-- anymore as it is deprecated.
+-- @deprecated
+-- @hook
+-- @realm client
+function GM:TTTSettingsTabs()
+
+end