+++ gamemodes/terrortown/gamemode/shared/sh_voice.lua
@@ -2,8 +2,9 @@
 -- @section voice_manager
 
 ---
--- Whether or not the @{Player} use the voice chat. Hint: Has to be registered
--- on both client and server to hide the UI element and stop the voicechat
+-- Whether or not the @{Player} can use the voice chat. 
+-- @note Has to be registered on both client and server to hide the UI
+-- element and stop the voicechat
 -- @param Player ply @{Player} who wants to use the voice chat
 -- @param boolean isTeam Are they trying to use the team voice chat
 -- @return[default=true] boolean Whether or not the @{Player} can use the voice chat