+++ lua/ttt2/libraries/outline.lua
@@ -112,8 +112,8 @@ local function Render()
 			local ent = ents[j]
 
 			if not IsValid(ent)
-			or mode == OUTLINE_MODE_NOTVISIBLE and IsLineOfSightClear(ent)
-			or mode == OUTLINE_MODE_VISIBLE and not IsLineOfSightClear(ent) then
+			or mode == OUTLINE_MODE_NOTVISIBLE and IsLineOfSightClear(client, ent)
+			or mode == OUTLINE_MODE_VISIBLE and not IsLineOfSightClear(client, ent) then
 				continue
 			end