+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ## Unreleased
 
+## [v0.8.2b](https://github.com/TTT-2/TTT2/tree/v0.8.1b) (2021-03-25)
+
 ### Fixed
 
 - TTT: fix instant reload of dropped weapon (by @svdm)
@@ -16,7 +18,6 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Updated Japanese localization (by @Westoon)
 - Moved rendering modules to libraries
 - Assigned PvP category to the gamemode.
-
 
 ## [v0.8.1b](https://github.com/TTT-2/TTT2/tree/v0.8.1b) (2021-02-19)